requirements.txt
pytest==8.1.1
jupyter==1.0.0
aiohttp==3.10.10
asyncpg==0.30.0
dbt-core==1.8.8
dbt-postgres==1.8.2
schedule==1.2.2
pandas==2.2.3
SQLAlchemy==2.0.36
matplotlib==3.9.2
